add user-listing relations, views, seeding and manage lissting func

+++ app/Http/Controllers/ListingController.php
@@ -42,7 +42,7 @@ class ListingController extends Controller
         if ($request->hasFile('logo')) {
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
-
+        $formFields['user_id'] = auth()->id();
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing created successfully!');
@@ -55,6 +55,11 @@ class ListingController extends Controller
     
     //Update Listing Data
     public function update(Request $request, Listing $listing) {
+        //Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action!');
+        }
+        
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -68,7 +73,7 @@ class ListingController extends Controller
         if ($request->hasFile('logo')) {
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
-
+        
         $listing->update($formFields);
 
         return back()->with('message', 'Listing updated successfully!');
@@ -76,7 +81,16 @@ class ListingController extends Controller
     
     //Delete listing
     public function destroy(Listing $listing) {
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action!');
+        }
+        
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully!');
+    }
+    
+    // Manage Listings
+    public function manage() {
+       // return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
